+++ web/internal/canvas_context_details_view.css
@@ -1,0 +1,85 @@
+/* Copyright 2026 Mozilla Foundation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+.gfx-state-section {
+  padding-inline: 12px;
+}
+
+.gfx-state-section + .gfx-state-section {
+  margin-top: 8px;
+  padding-top: 8px;
+  border-top: 1px solid var(--border-color);
+}
+
+.gfx-state-title {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 4px;
+  color: var(--accent-color);
+  font-weight: bold;
+  margin-bottom: 4px;
+}
+
+.gfx-state-stack-nav {
+  display: flex;
+  align-items: center;
+  gap: 2px;
+  font-weight: normal;
+  font-size: 0.8em;
+}
+
+.gfx-state-stack-button {
+  padding: 0 3px;
+  border: 1px solid currentcolor;
+  border-radius: 2px;
+  background: transparent;
+  color: inherit;
+  cursor: pointer;
+  line-height: 1.3;
+
+  &:disabled {
+    cursor: default;
+    opacity: 0.35;
+  }
+}
+
+.gfx-state-stack-pos {
+  min-width: 4ch;
+  text-align: center;
+  font-variant-numeric: tabular-nums;
+}
+
+.gfx-state-row {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  padding: 1px 0;
+}
+
+.gfx-state-key {
+  color: var(--muted-color);
+  flex-shrink: 0;
+  min-width: 20ch;
+}
+
+.gfx-state-val {
+  color: var(--number-color);
+  flex: 1 1 0;
+  min-width: 0;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
