+++ web/internal/debugger.html
@@ -19,7 +19,7 @@ limitations under the License.
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <title>PDF.js — Debugging tools</title>
-    <link rel="stylesheet" href="pdf_internal_viewer.css" />
+    <link rel="stylesheet" href="debugger.css" />
   </head>
   <body>
     <div id="header">
@@ -34,8 +34,8 @@ limitations under the License.
       <span id="goto-input-hint" class="sr-only">
         Enter a page number (e.g. 5), a reference as numR (e.g. 10R) or numRgen (e.g. 10R2). Press Enter to navigate.
       </span>
-      <button id="debug-btn" hidden>Debug page</button>
-      <button id="debug-back-btn" hidden>← Back to tree</button>
+      <button id="debug-button" hidden>Debug page</button>
+      <button id="debug-back-button" hidden>← Back to tree</button>
       <span id="status" role="status" aria-live="polite"> Select a PDF file to explore its internal structure. </span>
       <a id="github-link" href="https://github.com/mozilla/pdf.js" target="_blank" rel="noopener noreferrer" aria-label="PDF.js on GitHub">
         <svg viewBox="0 0 16 16" aria-hidden="true" focusable="false">
@@ -49,25 +49,20 @@ limitations under the License.
     <div id="debug-view" hidden>
       <div id="render-panels">
         <div id="op-left-col">
-          <div id="op-top-row">
-            <div id="op-list-panel">
-              <div id="op-list" role="listbox" aria-label="Operator list"></div>
-            </div>
-            <div id="op-gfx-state-resizer" role="separator" aria-orientation="vertical" tabindex="0" hidden></div>
-            <div id="gfx-state-panel" aria-label="Graphics state" hidden></div>
+          <div id="op-list-panel">
+            <div id="op-list" role="listbox" aria-label="Operator list"></div>
           </div>
-          <div id="op-resizer" role="separator" aria-orientation="horizontal" tabindex="0"></div>
           <div id="op-detail-panel"></div>
+          <div id="gfx-state-panel" aria-label="Graphics state" hidden></div>
         </div>
-        <div id="render-resizer" role="separator" aria-orientation="vertical" tabindex="0"></div>
         <div id="canvas-panel">
           <div id="canvas-toolbar" role="toolbar" aria-label="Zoom controls">
-            <button id="zoom-out-btn" aria-label="Zoom out">−</button>
+            <button id="zoom-out-button" aria-label="Zoom out">−</button>
             <span id="zoom-level" aria-live="polite"></span>
-            <button id="zoom-in-btn" aria-label="Zoom in">+</button>
-            <button id="redraw-btn" aria-label="Redraw page">Redraw</button>
-            <button id="step-btn" aria-label="Step one instruction" disabled><u>S</u>tep</button>
-            <button id="continue-btn" aria-label="Continue to next breakpoint" disabled><u>C</u>ontinue</button>
+            <button id="zoom-in-button" aria-label="Zoom in">+</button>
+            <button id="redraw-button" aria-label="Redraw page">Redraw</button>
+            <button id="step-button" aria-label="Step one instruction" disabled><u>S</u>tep</button>
+            <button id="continue-button" aria-label="Continue to next breakpoint" disabled><u>C</u>ontinue</button>
           </div>
           <div id="canvas-scroll">
             <div id="canvas-wrapper">
@@ -93,25 +88,25 @@ limitations under the License.
 
     <!--#if GENERIC-->
     <!--<script src="../build/pdf.mjs" type="module"></script>-->
-    <!--<script src="pdf_internal_viewer.mjs" type="module"></script>-->
+    <!--<script src="debugger.mjs" type="module"></script>-->
     <!--#else-->
     <script type="importmap">
       {
         "imports": {
-          "pdfjs/": "../src/",
-          "pdfjs-lib": "../src/pdf.js",
+          "pdfjs/": "../../src/",
+          "pdfjs-lib": "../../src/pdf.js",
 
-          "display-cmap_reader_factory": "../src/display/cmap_reader_factory.js",
-          "display-standard_fontdata_factory": "../src/display/standard_fontdata_factory.js",
-          "display-wasm_factory": "../src/display/wasm_factory.js",
-          "display-fetch_stream": "../src/display/fetch_stream.js",
-          "display-network": "../src/display/network.js",
-          "display-node_stream": "../src/display/stubs.js",
-          "display-node_utils": "../src/display/stubs.js"
+          "display-cmap_reader_factory": "../../src/display/cmap_reader_factory.js",
+          "display-standard_fontdata_factory": "../../src/display/standard_fontdata_factory.js",
+          "display-wasm_factory": "../../src/display/wasm_factory.js",
+          "display-fetch_stream": "../../src/display/fetch_stream.js",
+          "display-network": "../../src/display/network.js",
+          "display-node_stream": "../../src/display/stubs.js",
+          "display-node_utils": "../../src/display/stubs.js"
         }
       }
     </script>
-    <script src="pdf_internal_viewer.js" type="module"></script>
+    <script src="debugger.js" type="module"></script>
     <!--#endif-->
   </body>
 </html>
