+++ web/internal/split_view.css
@@ -1,0 +1,71 @@
+/* Copyright 2026 Mozilla Foundation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/* Hide the resizer automatically when the adjacent panel is not visible. */
+.spc-container > .spc-resizer:has(+ [hidden]),
+.spc-container > [hidden] + .spc-resizer {
+  display: none;
+}
+
+.spc-container {
+  display: flex;
+  overflow: hidden;
+
+  > * {
+    min-width: 0;
+    min-height: 0;
+  }
+
+  > .spc-resizer {
+    flex-shrink: 0;
+    background: var(--spc-resizer-color, #ccc);
+
+    &:hover,
+    &.dragging {
+      background: var(--spc-resizer-hover-color, #888);
+    }
+  }
+
+  &.spc-row {
+    flex-direction: row;
+
+    > .spc-resizer {
+      width: 6px;
+      cursor: col-resize;
+      align-self: stretch;
+    }
+  }
+
+  &.spc-column {
+    flex-direction: column;
+
+    > .spc-resizer {
+      height: 6px;
+      cursor: row-resize;
+      align-self: stretch;
+    }
+  }
+}
+
+@media (forced-colors: active) {
+  .spc-container > .spc-resizer {
+    background: ButtonBorder;
+
+    &:hover,
+    &.dragging {
+      background: Highlight;
+    }
+  }
+}
